// src/core/models/nrg-dtos/WorkOrderDto.ts
export interface WorkOrderDto {
  ProjectNumber: string
  ProjectName: string
  Id: string
  Number: string
  Name: string
  Type: string
  // CreatedBy: CreatedBy
  // CreatedOn: string
  // Facility: string
  // Outsourced: boolean
  // Tags: string[]
  Status: string
  // MaterialOnHandDays: number
  Step: string
  // StepIndex: number
  StepType: string
  // InvoiceStatus: string
  Owner?: Person
  Assignees?: Person[]
  // Drafters?: Drafter[]
  // Engineers?: Engineer[]
  // Estimators?: Estimator[]
  // SalesPersons?: SalesPerson[]
  // Coordinators?: Coordinator[]
  // Installers?: Installer[]
  ProjectManager?: Person
  // PlannedStartDate: string
  // ActualStartDate?: string
  // PlannedCriticalDate: string
  // MaterialNeededDate?: string
  // PlannedEndDate: string
  // PlannedEndMonth: string
  // ActualEndDate?: string
  // ActualEndMonth?: string
  Instructions: string
  // EstimatedLaborCost: EstimatedLaborCost
  // EstimatedMaterialCost: EstimatedMaterialCost
  // EstimatedCost: EstimatedCost
  // EstimatedHours: string
  // EstimatedMargin: EstimatedMargin
  // LaborHours: string
  // RemainingHours: string
  // PlannedHours: string
  // LaborCost: LaborCost
  // PlannedLaborCost: PlannedLaborCost
  // LaborGrandTotalPrice: LaborGrandTotalPrice
  // ActualLaborHours: string
  // ActualCost: ActualCost
  // ActualMaterialCost: ActualMaterialCost
  // ActualLaborCost: ActualLaborCost
  // ActualExpensesCost: ActualExpensesCost
  // ActualMargin: ActualMargin
  // MarginVariance: MarginVariance
  // GrandTotalPrice: GrandTotalPrice
  // PreSalesTaxPrice: PreSalesTaxPrice
  // SalesTax: SalesTax
  // ExternalIdentifier: string
  // ProductsQuantitySum: number
  // ShipmentItemsQuantitySum: number
  // ReadyForFieldVerificationStepItemPlannedDate?: string
  // ReadyForFieldVerificationStepItemActualDate?: string
  // FieldVerifiedStepItemPlannedDate?: string
  // FieldVerifiedStepItemActualDate?: string
  // PlannedLaborLastUpdatedDate?: string
  // PercentInstalled: number
  // PercentInstalledLastUpdatedDate?: string
  // ProjectedInstallRemainingHours: string
  // ShipmentDate?: string
  // PlannedShipmentDate?: string
  // ActualShipmentDate?: string
  // JobCostingStatus: string
  // TeamLead?: TeamLead
  // Impediments?: string[]
  // CustomFields: CustomField[]
  MasterNote?: string
  // WorkflowName: string
  // WorkflowStepEnteredDate: string
  // WorkflowStepEnteredDays: number
}

// export interface CreatedBy {
//   Id: string
//   FullName: string
// }

export interface Person {
  Id: string
  FullName: string
}